Dead code + FunctionalCharacterization split-up
