+++ .prettierignore
@@ -1,0 +1,2 @@
+# graphql-codegen output — formatting is owned by the generator, not by us
+src/common/types/generated/
