+++ doctor.config.json
@@ -1,0 +1,5 @@
+{
+  "ignore": {
+    "files": ["src/common/types/generated/**"]
+  }
+}
