+++ src/common/colors.ts
@@ -1,49 +1,10 @@
-export const promoterRed = "red";
-export const enhancerYellow = "#E9D31C";
-
-// strand
-export const geneRed = "#FF0000";
-export const geneBlue = "#1E90FF";
-
 /**
- * ENCODE Data
+ * Assay Colors
  */
-export const DNase_seq = "#06DA93";
-export const ATAC_seq = "#02c7b9";
-
-export const totalRNA_seq = "#00aa00";
-export const longreadRNA_seq = "#006600";
-
-// Histone mark
-export const H3K4me1 = "#FFDF00";
+export const DNase = "#06DA93";
 export const H3K4me3 = "#FF0000";
 export const H3K27ac = "#FFCD00";
-export const H3K36me3 = "#008000";
-export const H3K9me3 = "#B4DDE4";
-export const H3K27me3 = "#AEAFAE";
-
-// ChIA_PET
-export const ChIA_PETRNAPII = "#b20600";
-export const ChIA_PETCTCF = "#1900a6";
-export const Hi_C = "#9222b0";
-
-// Bru_seq
-export const Bru_seq = "#D642CA";
-export const BruUV_seq = "";
-export const BruChase_seq = "";
-
-// Pro
-export const PRO_cap = "#f68800";
-export const PRO_seq = "#b45f06";
-
-// other
-export const RAMPAGEPeaksSignal = "#D642CA";
-export const Conservation = "#999999";
-export const Genes = "#000000";
-export const TFChIP_seq = "#1262EB";
-export const RBP = "#f68800";
-
-export const Enhancer_GeneLinks = "#A872E5";
+export const CTCF = "#00B0F0";
 
 /**
  * MOHD Project
