+++ src/common/components/EQTLTables.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useQuery } from "@apollo/client/react";
 import { Stack } from "@mui/material";
-import { toScientificNotationElement } from "common/utils";
+import { ScientificNotation } from "common/utils";
 import { gql } from "common/types/generated";
 import { LinkComponent } from "./LinkComponent";
 import { TableColDef, Table } from "@weng-lab/ui-components";
@@ -109,13 +109,13 @@ export default function EQTLs({ entity }: EntityViewComponentProps) {
       field: "slope",
       headerName: "Slope",
       display: "flex",
-      renderCell: (params) => toScientificNotationElement(params.value, 2, { variant: "body2" }),
+      renderCell: (params) => ScientificNotation(params.value, 2, { variant: "body2" }),
     },
     {
       field: "pval_nominal",
       headerName: "Q Value",
       display: "flex",
-      renderCell: (params) => toScientificNotationElement(params.value, 2, { variant: "body2" }),
+      renderCell: (params) => ScientificNotation(params.value, 2, { variant: "body2" }),
     },
     {
       field: "celltype",
@@ -185,13 +185,13 @@ export default function EQTLs({ entity }: EntityViewComponentProps) {
       field: "fdr",
       headerName: "FDR",
       display: "flex",
-      renderCell: (params) => toScientificNotationElement(params.value, 2, { variant: "body2" }),
+      renderCell: (params) => ScientificNotation(params.value, 2, { variant: "body2" }),
     },
     {
       field: "spearmans_rho",
       headerName: "Spearman's rho",
       display: "flex",
-      renderCell: (params) => toScientificNotationElement(params.value, 2, { variant: "body2" }),
+      renderCell: (params) => ScientificNotation(params.value, 2, { variant: "body2" }),
     },
     {
       field: "celltype",
