+++ src/common/utils/scientificNotation.tsx
@@ -1,40 +1,12 @@
 import { Typography, TypographyOwnProps } from "@mui/material";
 
 /**
- * Only use this function if use case is unable to handle jsx element and needs string. Use `toScientificNotationElement` if possible
- */
-export function toScientificNotation(num: number, sigFigs?: number) {
-  // Convert the number to scientific notation using toExponential
-  const scientific = num.toExponential(sigFigs ?? undefined);
-
-  // Split the scientific notation into the coefficient and exponent parts
-
-  let [coefficient, exponent] = scientific.split("e");
-
-  // Format the exponent part
-  const expSign = exponent[0];
-  exponent = exponent.slice(1);
-
-  // Convert the exponent to a superscript string
-  let superscriptExponent = exponent
-    .split("")
-    .map((char) => "⁰¹²³⁴⁵⁶⁷⁸⁹"[char] || char)
-    .join("");
-
-  // Add the sign back to the exponent
-  superscriptExponent = (expSign === "-" ? "⁻" : "") + superscriptExponent;
-
-  // Combine the coefficient with the superscript exponent
-  return coefficient + "×10" + superscriptExponent;
-}
-
-/**
  * @param num Number to convert to Sci Notation
  * @param sigFigs Number of desired significant figures
  * @param typographyProps Props spread onto Typography element
  * @returns
  */
-export function toScientificNotationElement(num: number, sigFigs: number, typographyProps?: TypographyOwnProps) {
+export function ScientificNotation(num: number, sigFigs: number, typographyProps?: TypographyOwnProps) {
   if (num > 0.01) {
     return <Typography {...typographyProps}>{num.toFixed(2)}</Typography>;
   }
